airflow/requirements.txt
dbt-snowflake
boto3
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
astronomer-cosmos[dbt-snowflake]
